+++ src/pages/ManageAccountPage/components/AccountRowEmpty.jsx
@@ -1,0 +1,13 @@
+const AccountRowEmpty = () => {
+    return (
+        <thead className="text-md border-b-2 shadow-sm ">
+            <tr>
+                <th className="py-7 text-center font-medium text-gray-400 tracking-wider" scope="col" colSpan="9">
+                    ไม่มีข้อมูล
+                </th>
+            </tr>
+        </thead>
+    );
+};
+
+export default AccountRowEmpty;
